Refactors day 04
