+++ challenges/day_03/tests/test_challenges_day_03.py
