+++ lib/password_solver/password_solver.py
